MOONWORLD: Disabled object pie menu's except Touch and Sit.
Approach is the same as with normal menu's:
The default value of the 'moonworld' attribute (when it
is not there) is 'true', causing everything to be added as before.

By adding the 'moonworld' attribute, the default for that
node's children becomes 'false'.

If the attribute is true, the node is added, if it is false
(or absent, and the parent has the attribute) then the node
and all it's children are skipped.

This allows to restrict the pie menu's with only a minimal
amount of changes to it's xml file and makes it flexible
to enable/disable things later.