MOONWORLD: Remove toolbar; move snapshot and inventory buttons to overlay bar.
This is a bit of a hack cause I used the existing "windlight_remote"
panel (with the old Editor and Sky buttons) for the new place
of the Snapshot and Inventory buttons. So, the names of the xml
files and the names of the panels still contain "windlight", but
now deal with the Snapshot and Inventory button.

The whole patch took me 7 hours though :( ... So... um, it works.